# Conflicts:
#	FFXIVClassic Lobby Server/Database.cs
#	FFXIVClassic Map Server/Database.cs
#	FFXIVClassic Map Server/FFXIVClassic Map Server.csproj
#	FFXIVClassic Map Server/actors/chara/player/Inventory.cs
#	FFXIVClassic Map Server/actors/chara/player/Player.cs
#	FFXIVClassic Map Server/dataobjects/Session.cs
#	FFXIVClassic World Server/Server.cs

+++ FFXIVClassic Map Server/actors/chara/BattleSave.cs
@@ -5,7 +5,7 @@
         public float potencial = 6.6f;
         public short[] skillLevel = new short[52];
         public short[] skillLevelCap = new short[52];
-        public short[] skillPoint = new short[52];
+        public int[] skillPoint = new int[52];
 
         public short physicalLevel;
         public int physicalExp;
